# feat: add concurrency option to parallelize package loading