This PR adds support for running package loading in parallel due to `gosec.load` being a major bottleneck for big projects. 

The concurrency option defaults to the number of available CPUs and tests are using a concurrency value of 1.

## Benchmarks:

Benchmarks of a couple of GitLab projects using [hyperfine](https://github.com/sharkdp/hyperfine/):
___

### [Gitlab Pages](https://gitlab.com/gitlab-org/gitlab-pages):

#### before:

```
Benchmark 1: /path/to/gosec -exclude-generated ./...
Time (mean ± σ):     11.384 s ±  0.227 s    [User: 28.443 s, System: 8.731 s]
Range (min … max):   10.974 s … 11.793 s    10 runs

Warning: Ignoring non-zero exit code.
```

#### after:

```
Benchmark 1: /path/to/gosec -exclude-generated ./...
Time (mean ± σ):      2.292 s ±  0.078 s    [User: 23.965 s, System: 8.732 s]
Range (min … max):    2.230 s …  2.469 s    10 runs

Warning: Ignoring non-zero exit code.
```

### [Gitaly](https://gitlab.com/gitlab-org/gitaly/)

#### before:

```
Benchmark 1: /path/to/gosec -exclude-generated ./...
Time (mean ± σ):     27.442 s ±  0.710 s    [User: 75.289 s, System: 20.270 s]
Range (min … max):   26.601 s … 28.497 s    10 runs

Warning: Ignoring non-zero exit code.
```

#### after:

```
Benchmark 1: /path/to/gosec -exclude-generated ./...
Time (mean ± σ):      5.953 s ±  1.047 s    [User: 67.921 s, System: 20.703 s]
Range (min … max):    5.237 s …  8.658 s    10 runs

Warning: Ignoring non-zero exit code.
```