@ccojocar 2022-02-16
@kruskall Thanks for this improvement. I'll have a closer look at it soon. Please could you add some unit tests where the concurrency is more than `1`. I'll like to see that the concurrent case is covered by test.