@ccojocar 2022-02-16
`Check` cannot be made concurrent due to the shared state in the analyser.