@kruskall 2022-02-16
> @kruskall Thanks for this improvement. I'll have a closer look at it soon. Please could you add some unit tests where the concurrency is more than `1`. I'll like to see that the concurrent case is covered by test.

@ccojocar Thank you for the feedback! I've added a test with a concurrency value of 32, a few more comments and made sure a goroutine would stop if there are no jobs left. Back to you :ping_pong: